+++ src/main/java/ua/com/foxminded/domain/service/impl/TeacherServiceImpl.java
@@ -1,5 +1,6 @@
 package ua.com.foxminded.domain.service.impl;
 
+import javassist.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -10,6 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import ua.com.foxminded.dao.TeacherRepository;
 import ua.com.foxminded.domain.entity.Teacher;
 import ua.com.foxminded.domain.service.TeacherService;
+import ua.com.foxminded.exception.AlreadyExistException;
 
 import java.util.List;
 
@@ -25,29 +27,65 @@ public class TeacherServiceImpl implements TeacherService {
     }
 
     @Override
-    public void create(Teacher teacher) {
+    public void create(Teacher teacher) throws AlreadyExistException {
+        if (teacherRepository.findByNameOrSurname(teacher.getName()) != null) {
+            throw new AlreadyExistException("Teacher with the same name already exists");
+        }
         teacherRepository.save(teacher);
     }
 
     @Override
-    public void delete(Long id) {
+    public void update(Teacher teacher) throws NotFoundException {
+        if (teacherRepository.findByNameOrSurname(teacher.getName()) == null) {
+            throw new NotFoundException("Teacher with that name does not exists");
+        }
+        teacherRepository.save(teacher);
+    }
+
+    @Override
+    public void delete(Long id) throws NotFoundException {
+        Teacher teacher = teacherRepository.findById(id).get();
+        if (teacher == null) {
+            throw new NotFoundException("Teacher could not be found");
+        }
         teacherRepository.deleteById(id);
     }
 
     @Override
-    public Page<Teacher> findAll(int pageNumber) {
+    public Page<Teacher> findAll(int pageNumber) throws NotFoundException {
         Pageable pageable = PageRequest.of(pageNumber - 1, 10, Sort.by("name"));
-        return teacherRepository.findAll(pageable);
+        Page<Teacher> teachers = teacherRepository.findAll(pageable);
+        if (teachers == null) {
+            throw new NotFoundException("Page with teachers could not be found");
+        }
+        return teachers;
     }
 
     @Override
-    public Teacher findById(Long id) {
-        return teacherRepository.getOne(id);
+    public Teacher findById(Long id) throws NotFoundException {
+        Teacher teacher = teacherRepository.findById(id).get();
+        if (teacher == null) {
+            throw new NotFoundException("Teacher could not be found");
+        }
+        return teacher;
     }
 
     @Override
-    public List<Teacher> findByPersonalData(String name) {
-        return teacherRepository.findByNameOrSurname(name);
+    public List<Teacher> findByPersonalData(String name) throws NotFoundException {
+        List<Teacher> teachers = teacherRepository.findByNameOrSurname(name);
+        if (teachers == null) {
+            throw new NotFoundException("Teachers could not be found");
+        }
+        return teachers;
 
+    }
+
+    @Override
+    public List<Teacher> findAll() throws NotFoundException {
+        List<Teacher> teachers = teacherRepository.findAll();
+        if (teachers == null) {
+            throw new NotFoundException("Teachers could not be  found");
+        }
+        return teachers;
     }
 }
